# docs/api-reference/openapi.yaml
openapi: 3.0.3
info:
  title: Retail POS SaaS API
  version: 1.0.0
  description: Multi-tenant backend for fashion retail point of sale.
paths:
...
  /api/v1/purchases/:
    get:
      operationId: purchases_list
      description: |-
        Purchases are created and received, never edited in place.

        Editing a received purchase would silently rewrite the origin of stock the
        ledger already recorded.
      parameters:
      - in: query
        name: location
        schema:
          type: string
          format: uuid
      - name: ordering
        required: false
        in: query
        description: Qué campo usar para ordenar los resultados.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      - in: query
        name: purchased_after
        schema:
          type: string
          format: date-time
      - in: query
        name: purchased_before
        schema:
          type: string
          format: date-time
      - name: search
        required: false
        in: query
        description: Un término de búsqueda.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
          enum:
          - CANCELLED
          - DRAFT
          - RECEIVED
        description: |-
          * `DRAFT` - Draft
          * `RECEIVED` - Received
          * `CANCELLED` - Cancelled
      - in: query
        name: supplier
        schema:
          type: string
          format: uuid
      tags:
      - purchases
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPurchaseList'
          description: ''
    post:
      operationId: purchases_create
      description: |-
        Purchases are created and received, never edited in place.

        Editing a received purchase would silently rewrite the origin of stock the
        ledger already recorded.
      tags:
      - purchases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PurchaseCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PurchaseCreateRequest'
        required: true
      security:
      - organizationJWT: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
          description: ''
...
  /api/v1/reports/cash-sessions/:
    get:
      operationId: reports_cash_sessions_retrieve
      description: Read-only aggregates. Every figure is derived, never stored.
      parameters:
      - in: query
        name: from
        schema:
          type: string
        description: ISO-8601 start. Defaults to 30 days ago.
      - in: query
        name: location
        schema:
          type: string
        description: Restrict to one location.
      - in: query
        name: to
        schema:
          type: string
        description: ISO-8601 end. Defaults to now.
      tags:
      - reports
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashSessionsReport'
          description: ''
  /api/v1/reports/inventory-valuation/:
    get:
      operationId: reports_inventory_valuation_retrieve
      description: Read-only aggregates. Every figure is derived, never stored.
      parameters:
      - in: query
        name: location
        schema:
          type: string
      tags:
      - reports
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryValuation'
          description: ''
...
  /api/v1/reports/sales-summary/:
    get:
      operationId: reports_sales_summary_retrieve
      description: Read-only aggregates. Every figure is derived, never stored.
      parameters:
      - in: query
        name: from
        schema:
          type: string
        description: ISO-8601 start. Defaults to 30 days ago.
      - in: query
        name: location
        schema:
          type: string
        description: Restrict to one location.
      - in: query
        name: to
        schema:
          type: string
        description: ISO-8601 end. Defaults to now.
      tags:
      - reports
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesSummary'
          description: ''
  /api/v1/reports/top-products/:
    get:
      operationId: reports_top_products_list
      description: Read-only aggregates. Every figure is derived, never stored.
      parameters:
      - in: query
        name: from
        schema:
          type: string
        description: ISO-8601 start. Defaults to 30 days ago.
      - in: query
        name: limit
        schema:
          type: integer
        description: Default 10.
      - in: query
        name: location
        schema:
          type: string
        description: Restrict to one location.
      - in: query
        name: to
        schema:
          type: string
        description: ISO-8601 end. Defaults to now.
      tags:
      - reports
      security:
      - organizationJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopProduct'
          description: ''
...
components:
  schemas:
...
    CashSessionRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        register:
          type: string
        opened_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        expected:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        counted:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        difference:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        closed_by:
          type: string
          nullable: true
      required:
      - closed_at
      - closed_by
      - counted
      - difference
      - expected
      - id
      - opened_at
      - register
...
    CashSessionsReport:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/Period'
        sessions_closed:
          type: integer
        total_shortfall:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        total_surplus:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/CashSessionRow'
      required:
      - period
      - sessions
      - sessions_closed
      - total_shortfall
      - total_surplus
...
    DailySales:
      type: object
      properties:
        date:
          type: string
          format: date
        sales_count:
          type: integer
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - date
      - sales_count
      - total
...
    InventoryValuation:
      type: object
      properties:
        units_on_hand:
          type: integer
        cost_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        retail_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        potential_margin:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        variants_tracked:
          type: integer
        negative_stock:
          type: array
          items:
            $ref: '#/components/schemas/NegativeStock'
      required:
      - cost_value
      - negative_stock
      - potential_margin
      - retail_value
      - units_on_hand
      - variants_tracked
...
    NegativeStock:
      type: object
      properties:
        variant:
          type: string
          format: uuid
        sku:
          type: string
        location:
          type: string
        quantity:
          type: integer
      required:
      - location
      - quantity
      - sku
      - variant
...
    PaginatedPurchaseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Purchase'
...
    Period:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required:
      - end
      - start
...
    Purchase:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: string
          readOnly: true
        location:
          type: string
          format: uuid
          readOnly: true
        location_name:
          type: string
          readOnly: true
        supplier:
          type: string
          format: uuid
          readOnly: true
        supplier_name:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/PurchaseStatusEnum'
          readOnly: true
        supplier_invoice:
          type: string
          readOnly: true
        purchased_at:
          type: string
          format: date-time
          readOnly: true
        received_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        total_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
        notes:
          type: string
          readOnly: true
        created_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        received_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseItem'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - created_by
      - id
      - items
      - location
      - location_name
      - notes
      - number
      - purchased_at
      - received_at
      - received_by
      - status
      - supplier
      - supplier_invoice
      - supplier_name
      - total_cost
    PurchaseCreateRequest:
      type: object
      properties:
        location:
          type: string
          format: uuid
        supplier:
          type: string
          format: uuid
        supplier_invoice:
          type: string
          maxLength: 40
        purchased_at:
          type: string
          format: date-time
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseItemInputRequest'
        receive:
          type: boolean
          default: false
          description: Receive into stock immediately. The common flow for a small
            store.
      required:
      - items
      - supplier
    PurchaseItem:
      type: object
      description: |-
        Base serializer for tenant-owned models.

        Overriding `serializer_related_field` makes every auto-generated relation
        tenant-scoped, so isolation does not depend on each developer remembering
        to scope a nested id by hand. `organization` is never writable: it comes
        from the authenticated context.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        variant:
          type: string
          format: uuid
        variant_sku:
          type: string
          readOnly: true
        variant_name:
          type: string
          readOnly: true
        quantity:
          type: integer
          maximum: 2147483647
          minimum: 1
        unit_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        total_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - id
      - quantity
      - total_cost
      - unit_cost
      - variant
      - variant_name
      - variant_sku
    PurchaseItemInputRequest:
      type: object
      properties:
        variant:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unit_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
      required:
      - quantity
      - unit_cost
      - variant
    PurchaseStatusEnum:
      enum:
      - DRAFT
      - RECEIVED
      - CANCELLED
      type: string
      description: |-
        * `DRAFT` - Draft
        * `RECEIVED` - Received
        * `CANCELLED` - Cancelled
...
    SalesSummary:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/Period'
        sales_count:
          type: integer
        gross_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        refunded_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        net_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        tax_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        discount_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        average_ticket:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
        payments_by_method:
          type: object
          additionalProperties:
            type: string
            format: decimal
            pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
            readOnly: true
        by_day:
          type: array
          items:
            $ref: '#/components/schemas/DailySales'
      required:
      - average_ticket
      - by_day
      - discount_total
      - gross_total
      - net_total
      - payments_by_method
      - period
      - refunded_total
      - sales_count
      - tax_total
...
    TopProduct:
      type: object
      properties:
        variant:
          type: string
          format: uuid
        sku:
          type: string
        description:
          type: string
        units:
          type: integer
        revenue:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - description
      - revenue
      - sku
      - units
      - variant
...
  securitySchemes:
    organizationJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /api/v1/auth/login/. An account belongs to exactly
        one business, so the tenant is derived from the account on every request -
        there is nothing to select and nothing to switch.
servers:
- url: http://localhost:8000
  description: Local development